# product/src/main/openapi/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    ProductDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          pattern: \S
          maxLength: 100
        unitPrice:
          type: number
        quantityTotal:
          type: number
        quantityReserved:
          type: number
      required:
      - name
      - unitPrice
      - quantityTotal
paths:
  /v1/product:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Add Product
      tags:
      - Product Resource
  /v1/product/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDto"
      summary: Find Product
      tags:
      - Product Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Delete Product
      tags:
      - Product Resource
  /v1/products:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductDto"
      summary: Find All Products
      tags:
      - Product Resource
info:
  title: product API
  version: 1.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value